Move child document components to separate files
